V1.0.0 (#68)
``django-river`` is finally having it's first major version bump. In this version, all code and the APIs are revisited
and are much easier to understand how it works and much easier to use it now. In some places even more performant. 
There are also more documentation with this version. Stay tuned :-)

    * Support ``Django2.1``
    * Support multiple state fields in a model
    * Make the API very easy and useful by accessing everyhing via model objects and model classes
    * Simplify the concepts
    * Migrate ProceedingMeta and Transition into TransitionApprovalMeta for simplification
    * Rename Proceeding as TransitionApproval
    * Document transition and on-complete hooks
    * Document transition and on-complete hooks
    * Imrove documents in general
    * Minor improvements on admin pages
    * Some performance improvements